move hashing to new file
